test: Support gcc-14.
* lib/tests/scaffold/30-exit-0.c,
lib/tests/scaffold/30-exit-42.c,
lib/tests/scaffold/48-global-static.c,
lib/tests/scaffold/51-pointer-sub.c,
lib/tests/scaffold/62-array.c,
lib/tests/scaffold/63-struct-cell.c,
lib/tests/scaffold/63-struct-function.c,
lib/tests/scaffold/64-make-cell.c,
lib/tests/scaffold/70-stdarg.c,
lib/tests/scaffold/71-struct-array.c,
lib/tests/scaffold/72-typedef-struct-def-local.c,
lib/tests/scaffold/72-typedef-struct-def.c,
lib/tests/scaffold/76-pointer-arithmetic.c,
lib/tests/scaffold/77-pointer-assign.c,
lib/tests/scaffold/79-int-array-simple.c,
lib/tests/scaffold/7a-struct-char-array.c,
lib/tests/scaffold/7b-struct-int-array-pointer.c,
lib/tests/scaffold/7b-struct-int-array.c,
lib/tests/scaffold/7c-dynarray.c,
lib/tests/scaffold/7f-struct-pointer-arithmetic.c,
lib/tests/scaffold/7k-for-each-elem.c,
lib/tests/scaffold/7l-struct-any-size-array.c,
lib/tests/scaffold/7n-struct-struct-array.c,
lib/tests/scaffold/7t-function-destruct.c,
lib/tests/scaffold/7u-struct-size10.c,
lib/tests/scaffold/85-sizeof.c,
lib/tests/scaffold/a1-global-no-align.c,
lib/tests/scaffold/t.c,
lib/tests/stdlib/50-malloc.c,
lib/tests/stdlib/70-strtoull.c,
lib/tests/stdlib/80-qsort-dupes.c,
lib/tests/stdlib/80-qsort.c,
lib/tests/string/90-strpbrk.c,
lib/tests/string/90-strspn.c: Add missing prototypes, missing includes,
missing casts, etc.

+++ lib/tests/stdlib/50-malloc.c
@@ -1,6 +1,6 @@
 /* -*-comment-start: "//";comment-end:""-*-
  * GNU Mes --- Maxwell Equations of Software
- * Copyright © 2016,2017 Jan (janneke) Nieuwenhuizen <janneke@gnu.org>
+ * Copyright © 2016,2017,2025 Janneke Nieuwenhuizen <janneke@gnu.org>
  *
  * This file is part of GNU Mes.
  *
@@ -31,12 +31,11 @@ main (int argc, char *argv[])
 {
   int size = 5000;
   oputs ("m!\n");
-  //int *p = 0;
   char *p = 0;
   p = malloc (size);
   size = 5000;
   oputs ("p=");
-  oputs (itoa (p));
+  oputs (itoa ((size_t)p));
   oputs ("\n");
   int i;
   for (i = 0; i < size; i = i + 1)
